Refactor: List 관련 비즈니스 로직 컴포넌트 내부로 넣기
