+++ src/features/post/ui/PostsFilter.tsx
@@ -6,37 +6,78 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/shared/ui';
+import { useLocation } from 'react-router-dom';
 import { Search } from 'lucide-react';
+import { useAtom, useAtomValue } from 'jotai';
+import {
+  skipAtom,
+  limitAtom,
+  searchQueryAtom,
+  sortByAtom,
+  sortOrderAtom,
+  selectedTagAtom,
+} from '@/features/post/model/atom';
+import { tagsAtom } from '@/entities/post/model/postAtoms';
+import { useSearchPosts } from '@/features/post/model/useSearchPosts';
+import { useFetchPostsByTag } from '../model/useFetchPostsByTag';
+import { useFetchPosts } from '@/features/post/model/useFetchPosts';
+import { useFetchTags } from '../model/useFetchTags';
+import { useEffect } from 'react';
+import { useUpdateURL } from '../model/useUpdateURL';
 
-interface PostFiltersProps {
-  searchQuery: string;
-  setSearchQuery: (value: string) => void;
-  searchPosts: () => void;
-  tags: { url: string; slug: string }[];
-  selectedTag: string;
-  setSelectedTag: (value: string) => void;
-  fetchPostsByTag: (tag: string) => void;
-  updateURL: () => void;
-  sortBy: string;
-  setSortBy: (value: string) => void;
-  sortOrder: 'asc' | 'desc';
-  setSortOrder: (value: 'asc' | 'desc') => void;
-}
+export const PostFilters = () => {
+  const location = useLocation();
+  const [skip, setSkip] = useAtom(skipAtom);
+  const [limit, setLimit] = useAtom(limitAtom);
+  const [searchQuery, setSearchQuery] = useAtom(searchQueryAtom);
+  const [sortBy, setSortBy] = useAtom(sortByAtom);
+  const [sortOrder, setSortOrder] = useAtom(sortOrderAtom);
+  const [selectedTag, setSelectedTag] = useAtom(selectedTagAtom);
+  const tags = useAtomValue(tagsAtom);
+  const { updateURL } = useUpdateURL();
+  const { fetchPostsByTag } = useFetchPostsByTag();
+  const { searchPosts } = useSearchPosts();
+  const { fetchTags } = useFetchTags();
+  const { fetchPosts } = useFetchPosts();
 
-export const PostFilters = ({
-  searchQuery,
-  setSearchQuery,
-  searchPosts,
-  tags,
-  selectedTag,
-  setSelectedTag,
-  fetchPostsByTag,
-  updateURL,
-  sortBy,
-  setSortBy,
-  sortOrder,
-  setSortOrder,
-}: PostFiltersProps) => {
+  useEffect(() => {
+    fetchTags();
+  }, [fetchTags]);
+
+  useEffect(() => {
+    if (selectedTag) fetchPostsByTag(selectedTag);
+    else fetchPosts();
+
+    updateURL();
+  }, [
+    skip,
+    limit,
+    sortBy,
+    sortOrder,
+    selectedTag,
+    updateURL,
+    fetchPostsByTag,
+    fetchPosts,
+  ]);
+
+  useEffect(() => {
+    const params = new URLSearchParams(location.search);
+    setSkip(parseInt(params.get('skip') || '0'));
+    setLimit(parseInt(params.get('limit') || '10'));
+    setSearchQuery(params.get('search') || '');
+    setSortBy(params.get('sortBy') || '');
+    setSortOrder((params.get('sortOrder') as 'asc' | 'desc') || 'asc');
+    setSelectedTag(params.get('tag') || '');
+  }, [
+    location.search,
+    setLimit,
+    setSkip,
+    setSearchQuery,
+    setSortBy,
+    setSortOrder,
+    setSelectedTag,
+  ]);
+
   return (
     <div className="flex gap-4">
       <div className="flex-1">
@@ -85,7 +126,10 @@ export const PostFilters = ({
         </SelectContent>
       </Select>
 
-      <Select value={sortOrder} onValueChange={setSortOrder}>
+      <Select
+        value={sortOrder}
+        onValueChange={value => setSortOrder(value as 'asc' | 'desc')}
+      >
         <SelectTrigger className="w-[180px]">
           <SelectValue placeholder="정렬 순서" />
         </SelectTrigger>
